+++ lib/src/iosMain/kotlin/com/paleblueapps/kmpcore/rating/RatingService.ios.kt
@@ -16,7 +16,7 @@ actual interface RatingService {
         ratingDialogConfig: DialogConfig,
         feedbackDialogConfig: DialogConfig,
         snoozeDuration: Duration,
-        minActionsNeededForAskingReview: Long,
+        minActionsNeededForAskingReview: Int,
     )
 
     actual suspend fun logUserAction()
